// Electron/370InsuranceApp-master/370InsuranceApp-master/src/app/Models/Farm.ts
export class Farm{
  constructor(){}
  Farm_ID:number;
  Farm_Type: string;
  Farm_Size:number;
  Farm_Name:string;
  Farm_Address: string;
  Farm_Email: string;
  Farm_ContactNumber: string;
  Farm_Number: string;
  Farm_Avg_Health_Score: number;
  Province_ID:number;
  Province_Description: string;
  Farm_Type_Description:string;
  Is_Active: string;
}